Attribute Macro lending::call_index

#[call_index]

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case Each dispatchable may also be annotated with the #[pallet::call_index($idx)] attribute, which explicitly defines the codec index for the dispatchable function in the Call enum.

All call indexes start from 0, until it encounters a dispatchable function with a defined call index. The dispatchable function that lexically follows the function with a defined call index will have that call index, but incremented by 1, e.g. if there are 3 dispatchable functions fn foo, fn bar and fn qux in that order, and only fn bar has a call index of 10, then fn qux will have an index of 11, instead of 1.

All arguments must implement [Debug], [PartialEq], [Eq], Decode, Encode, and [Clone]. For ease of use, bound by the trait frame_support::pallet_prelude::Member.

If no #[pallet::call] exists, then a default implementation corresponding to the following code is automatically generated:

#[pallet::call]
impl<T: Config> Pallet<T> {}

WARNING: modifying dispatchables, changing their order, removing some, etc., must be done with care. Indeed this will change the outer runtime call type (which is an enum with one variant per pallet), this outer runtime call can be stored on-chain (e.g. in pallet-scheduler). Thus migration might be needed. To mitigate against some of this, the #[pallet::call_index($idx)] attribute can be used to fix the order of the dispatchable so that the Call enum encoding does not change after modification. As a general rule of thumb, it is therefore adventageous to always add new calls to the end so you can maintain the existing order of calls.

Macro expansion

The macro creates an enum Call with one variant per dispatchable. This enum implements: [Clone], [Eq], [PartialEq], [Debug] (with stripped implementation in not("std")), Encode, Decode, GetDispatchInfo, GetCallName, GetCallIndex and UnfilteredDispatchable.

The macro implements the Callable trait on Pallet and a function call_functions which returns the dispatchable metadata.