Attribute Macro lending::generate_deposit

#[generate_deposit]

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case The attribute #[pallet::generate_deposit($visibility fn deposit_event)] generates a helper function on Pallet that handles deposit events.

NOTE: For instantiable pallets, the event must be generic over T and I.

Macro expansion

The macro will add on enum Event the attributes:

#[derive(frame_support::CloneNoBound)]

#[derive(frame_support::EqNoBound)]

#[derive(frame_support::PartialEqNoBound)]

#[derive(frame_support::RuntimeDebugNoBound)]

#[derive(codec::Encode)]

#[derive(codec::Decode)]

The macro implements From<Event<..>> for ().

The macro implements a metadata function on Event returning the EventMetadata.

If #[pallet::generate_deposit] is present then the macro implements fn deposit_event on Pallet.