Attribute Macro lending::hooks

#[hooks]

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case The #[pallet::hooks] attribute allows you to specify a Hooks implementation for Pallet that specifies pallet-specific logic.

The item the attribute attaches to must be defined as follows:

#[pallet::hooks]
impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> $optional_where_clause {
    ...
}

I.e. a regular trait implementation with generic bound: T: Config, for the trait Hooks<BlockNumberFor<T>> (they are defined in preludes), for the type Pallet<T> and with an optional where clause.

If no #[pallet::hooks] exists, then the following default implementation is automatically generated:

#[pallet::hooks]
impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {}

Macro expansion

The macro implements the traits OnInitialize, OnIdle, OnFinalize, OnRuntimeUpgrade, OffchainWorker, and IntegrityTest using the provided Hooks implementation.

NOTE: OnRuntimeUpgrade is implemented with Hooks::on_runtime_upgrade and some additional logic. E.g. logic to write the pallet version into storage.

NOTE: The macro also adds some tracing logic when implementing the above traits. The following hooks emit traces: on_initialize, on_finalize and on_runtime_upgrade.