Attribute Macro lending::import_section

#[import_section]

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case An attribute macro that can be attached to a module declaration. Doing so will Imports the contents of the specified external pallet section that was defined previously using #[pallet_section].

Example

#[import_section(some_section)]
#[pallet]
pub mod pallet {
    // ...
}

where some_section was defined elsewhere via:

#[pallet_section]
pub mod some_section {
    // ...
}

This will result in the contents of some_section being verbatim imported into the pallet above. Note that since the tokens for some_section are essentially copy-pasted into the target pallet, you cannot refer to imports that don’t also exist in the target pallet, but this is easily resolved by including all relevant use statements within your pallet section, so they are imported as well, or by otherwise ensuring that you have the same imports on the target pallet.

It is perfectly permissible to import multiple pallet sections into the same pallet, which can be done by having multiple #[import_section(something)] attributes attached to the pallet.

Note that sections are imported by their module name/ident, and should be referred to by their full path from the perspective of the target pallet.