Module lending::fungibles

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case The traits for sets of fungible tokens and any associated types.

Modules

approvals

Inspect and Mutate traits for Asset approvals

freeze

The traits for putting freezes within a single fungible token class.

hold

The traits for putting holds within a single fungible token class.

metadata

Inspect and Mutate traits for Asset metadata

roles

Inspect traits for Asset roles

Structs

DecreaseIssuance

Simple handler for an imbalance drop which decreases the total issuance of the system by the imbalance amount. Used for leftover credit.

Dust

Special dust type which can be type-safely converted into a Credit.

Imbalance

An imbalance in the system, representing a divergence of recorded token supply from the sum of the balances of all accounts. This is must_use in order to ensure it gets handled (placing into an account, settling from an account or altering the supply).

IncreaseIssuance

Simple handler for an imbalance drop which increases the total issuance of the system by the imbalance amount. Used for leftover debt.

Traits

Balanced

A fungible token class where any creation and deletion of tokens is semi-explicit and where the total supply is maintained automatically.

BalancedHold

Trait for slashing a fungible asset which can be place on hold.

Create

Trait for providing the ability to create new fungible assets.

Destroy

Trait for providing the ability to destroy existing fungible assets.

HandleImbalanceDrop

Handler for when an imbalance gets dropped. This could handle either a credit (negative) or debt (positive) imbalance.

Inspect

Trait for providing balance-inspection access to a set of named fungible assets.

InspectEnumerable

Interface for enumerating assets in existence or owned by a given account.

InspectFreeze

Trait for inspecting a fungible asset which can be frozen. Freezing is essentially setting a minimum balance below which the total balance (inclusive of any funds placed on hold) may not be normally allowed to drop. Generally, freezers will provide an “update” function such that if the total balance does drop below the limit, then the freezer can update their housekeeping accordingly.

InspectHold

Trait for inspecting a fungible asset whose accounts support partitioning and slashing.

Mutate

Trait for providing a basic fungible asset.

MutateFreeze

Trait for introducing, altering and removing locks to freeze an account’s funds so they never go below a set minimum.

MutateHold

Trait for mutating a fungible asset which can be placed on hold.

SwapNative

Trait for providing methods to swap between the chain’s native token and other asset classes.

Unbalanced

A fungible token class where the balance can be set arbitrarily.

UnbalancedHold

A fungible, holdable token class where the balance on hold can be set arbitrarily.

Type Aliases

Credit

Imbalance implying that the total_issuance value is greater than the sum of all account balances.

Debt

Imbalance implying that the total_issuance value is less than the sum of all account balances.