Trait lending::fungibles::MutateFreeze

pub trait MutateFreeze<AccountId>: Inspect<AccountId> {
    // Required methods
    fn set_freeze(
        asset: Self::AssetId,
        id: &Self::Id,
        who: &AccountId,
        amount: Self::Balance,
    ) -> Result<(), DispatchError>;
    fn extend_freeze(
        asset: Self::AssetId,
        id: &Self::Id,
        who: &AccountId,
        amount: Self::Balance,
    ) -> Result<(), DispatchError>;
    fn thaw(
        asset: Self::AssetId,
        id: &Self::Id,
        who: &AccountId,
    ) -> Result<(), DispatchError>;
}

Trait for introducing, altering and removing locks to freeze an account’s funds so they never go below a set minimum.

Required Methods

fn set_freeze( asset: Self::AssetId, id: &Self::Id, who: &AccountId, amount: Self::Balance, ) -> Result<(), DispatchError>

Prevent actions which would reduce the balance of the account of who below the given amount and identify this restriction though the given id. Unlike extend_freeze, any outstanding freeze in place for who under the id are dropped.

If amount is zero, it is equivalent to using thaw.

Note that amount can be greater than the total balance, if desired.

fn extend_freeze( asset: Self::AssetId, id: &Self::Id, who: &AccountId, amount: Self::Balance, ) -> Result<(), DispatchError>

Prevent the balance of the account of who from being reduced below the given amount and identify this restriction though the given id. Unlike set_freeze, this does not counteract any pre-existing freezes in place for who under the id. Also unlike set_freeze, in the case that amount is zero, this is no-op and never fails.

Note that more funds can be locked than the total balance, if desired.

fn thaw( asset: Self::AssetId, id: &Self::Id, who: &AccountId, ) -> Result<(), DispatchError>

Remove an existing lock.

Object Safety

This trait is not object safe.