Module lending::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

activate_lending_pool

The activate_lending_pool function allows a user to activate a lending pool that is not empty. Once a liquidity pool gets activated supplies operations can be performed otherwise only withdrawals.

borrow

The borrow function allows a user to borrow liquidity from a lending pool.

claim_rewards

The claim_rewards function allows a user to claim their rewards.

create_lending_pool

The create_lending_pool function allows a user to Create a new reserve and then supply it with some liquidity. Given an asset and its amount, it creates a new lending pool, if it does not already exist, and adds the provided liquidity

deactivate_lending_pool

The deactivate_lending_pool function allows a user to deactivate a lending pool that is not empty. Once a liquidity pool gets deactivated supplies operations can not be performed otherwise only withdrawals.

repay

The repay function allows a user to repay liquidity from a lending pool.

set_asset_price

Sets the price of one asset in terms of another asset.

supply

The supply function allows a user to supply liquidity to a lending pool.

update_pool_kink

The update_pool_kink function allows a user to update the kink of a lending pool.

update_pool_rate_model

The update_pool_rate_model function allows a user to update the rate model of a lending pool.

withdraw

The withdraw function allows a user to withdraw liquidity from a lending pool.