Function lending::pallet::dispatchables::create_lending_pool

pub fn create_lending_pool<T: Config>(
    id: LendingPoolId,
    asset: AssetIdOf<T>,
    balance: BalanceOf<T>,
)

The create_lending_pool function allows a user to Create a new reserve and then supply it with some liquidity. Given an asset and its amount, it creates a new lending pool, if it does not already exist, and adds the provided liquidity

The user will receive LP tokens in return in ratio.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

id: AssetIdOf - The pool id, provided by the user

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If amount is 0 or less.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers two events:

LendingPoolAdded(who, asset_a) if a new lending pool was created.

DepositSupplied(who, asset_a, amount_a) after the liquidity has been successfully added.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::create_lending_pool.