Function lending::pallet::dispatchables::set_asset_price

pub fn set_asset_price<T: Config>(
    asset_1: AssetIdOf<T>,
    asset_2: AssetIdOf<T>,
    price: FixedU128,
)

Sets the price of one asset in terms of another asset.

The set_asset_price extrinsic allows a user to specify the relative price of one asset (asset_1) in terms of another asset (asset_2).

Parameters

origin: The transaction origin. This must be a signed extrinsic.

asset_1: The identifier for the first asset. This is the asset whose price is being set.

asset_2: The identifier for the second asset. This is the asset relative to which the price is measured.

price: The price of asset_1 in terms of asset_2. This must be a non-zero value to avoid errors.

Events

AssetPriceAdded { asset_1, asset_2, price }: This event is emitted after the price is successfully set. It contains the asset identifiers and the new price.

Errors

InvalidAssetPrice: This error is thrown if the price parameter is zero.

Note this should be moved to a new pallet prices

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_asset_price.