Function lending::pallet::dispatchables::supply

pub fn supply<T: Config>(asset: AssetIdOf<T>, balance: BalanceOf<T>)

The supply function allows a user to supply liquidity to a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is not active.

If the user has not enough liquidity to supply.

If the balance amount to supply is not valid.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers an event:

DepositSupplied(who, asset, balance) if the lending pool has been supplied.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::supply.