Module lending::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AssetPrices

The storage to hold prices of assets w.r.t. other assets This is the dummy storage, ideally this functionality would be implemented in a dedicated pallet store (asset_id1, asset_id2) => FixedU128

Borrows

The borrow status of accounts (AccountId, borrowed_asset_id, collateral_asset_id) => UserBorrow details

LendingPoolStorage

Kylix runtime storage items

MinMaxExchangeRate

The minimum (starting) and maximum exchange rate allowed for a market.

SupplyIndexStorage

The accrued supply_index of accounts for assets

UnderlyingAssetStorage

Kylix runtime storage items