Struct lending::pallet::Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn create_lending_pool( origin: OriginFor<T>, id: LendingPoolId, asset: AssetIdOf<T>, balance: BalanceOf<T>, ) -> DispatchResult

The create_lending_pool function allows a user to Create a new reserve and then supply it with some liquidity. Given an asset and its amount, it creates a new lending pool, if it does not already exist, and adds the provided liquidity

The user will receive LP tokens in return in ratio.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

id: AssetIdOf - The pool id, provided by the user

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If amount is 0 or less.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers two events:

LendingPoolAdded(who, asset_a) if a new lending pool was created.

DepositSupplied(who, asset_a, amount_a) after the liquidity has been successfully added.

pub fn activate_lending_pool( origin: OriginFor<T>, asset: AssetIdOf<T>, ) -> DispatchResult

The activate_lending_pool function allows a user to activate a lending pool that is not empty. Once a liquidity pool gets activated supplies operations can be performed otherwise only withdrawals.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is already activated.

If the pool is empty.

Events

If the function succeeds, it triggers an event:

LendingPoolActivated(who, asset_a) if the lending pool was activated.

pub fn supply( origin: OriginFor<T>, asset: AssetIdOf<T>, balance: BalanceOf<T>, ) -> DispatchResult

The supply function allows a user to supply liquidity to a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is not active.

If the user has not enough liquidity to supply.

If the balance amount to supply is not valid.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers an event:

DepositSupplied(who, asset, balance) if the lending pool has been supplied.

pub fn withdraw( origin: OriginFor<T>, asset: AssetIdOf<T>, balance: BalanceOf<T>, ) -> DispatchResult

The withdraw function allows a user to withdraw liquidity from a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is not active.

If the user has not enough liquidity to supply.

If the balance amount to supply is not valid.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers an event:

DepositWithdrawn(who, balance) if the lending pool was activated.

pub fn borrow( origin: OriginFor<T>, asset: AssetIdOf<T>, balance: AssetBalanceOf<T>, collateral_asset: AssetIdOf<T>, collateral_balance: AssetBalanceOf<T>, ) -> DispatchResult

The borrow function allows a user to borrow liquidity from a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is not active.

If the user has not enough liquidity to supply.

If the balance amount to supply is not valid.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers an event:

DepositBorrowed(who, balance) if the lending pool was activated.

pub fn repay( origin: OriginFor<T>, asset: AssetIdOf<T>, balance: BalanceOf<T>, collateral_asset: AssetIdOf<T>, ) -> DispatchResult

The repay function allows a user to repay liquidity from a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

balance - The amount of asset that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is not active.

If the user has not enough liquidity to supply.

If the balance amount to supply is not valid.

If adding liquidity to the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers an event:

DepositRepaid(who, balance) if the lending pool was activated.

pub fn claim_rewards( origin: OriginFor<T>, balance: BalanceOf<T>, ) -> DispatchResult

The claim_rewards function allows a user to claim their rewards.

Arguments

origin - The origin caller of this function. This should be signed by the user claiming rewards.

balance - The amount of rewards to be claimed.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

Events

If the function succeeds, it triggers an event:

RewardsClaimed { who, balance }: Notifies the system that rewards have been claimed by a user.

pub fn deactivate_lending_pool( origin: OriginFor<T>, asset: AssetIdOf<T>, ) -> DispatchResult

The deactivate_lending_pool function allows a user to deactivate a lending pool that is not empty. Once a liquidity pool gets deactivated supplies operations can not be performed otherwise only withdrawals.

Arguments

origin - The origin caller of this function. This should be signed by the user that creates the lending pool and add some liquidity.

asset - The identifier for the type of asset that the user wants to provide.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If the pool does not exist.

If the pool is already deactivated.

If the pool is empty.

Events

If the function succeeds, it triggers an event:

LendingPoolDeactivated(who, asset_a) if the lending pool was deactivated.

pub fn update_pool_rate_model( origin: OriginFor<T>, asset: AssetIdOf<T>, ) -> DispatchResult

The update_pool_rate_model function allows a user to update the rate model of a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user updating the rate model.

asset - The identifier for the type of asset associated with the lending pool.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

Events

If the function succeeds, it triggers an event:

LendingPoolRateModelUpdated { who, asset }: Notifies the system that the rate model of a lending pool was updated by a user.

pub fn update_pool_kink( origin: OriginFor<T>, asset: AssetIdOf<T>, ) -> DispatchResult

The update_pool_kink function allows a user to update the kink of a lending pool.

Arguments

origin - The origin caller of this function. This should be signed by the user updating the kink.

asset - The identifier for the type of asset associated with the lending pool.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

Events

If the function succeeds, it triggers an event:

LendingPoolKinkUpdated { who, asset }: Notifies the system that the kink of a lending pool was updated by a user.

pub fn set_asset_price( origin: OriginFor<T>, asset_1: AssetIdOf<T>, asset_2: AssetIdOf<T>, price: FixedU128, ) -> DispatchResult

Sets the price of one asset in terms of another asset.

The set_asset_price extrinsic allows a user to specify the relative price of one asset (asset_1) in terms of another asset (asset_2).

Parameters

origin: The transaction origin. This must be a signed extrinsic.

asset_1: The identifier for the first asset. This is the asset whose price is being set.

asset_2: The identifier for the second asset. This is the asset relative to which the price is measured.

price: The price of asset_1 in terms of asset_2. This must be a non-zero value to avoid errors.

Events

AssetPriceAdded { asset_1, asset_2, price }: This event is emitted after the price is successfully set. It contains the asset identifiers and the new price.

Errors

InvalidAssetPrice: This error is thrown if the price parameter is zero.

Note this should be moved to a new pallet prices

impl<T: Config> Pallet<T>

pub fn do_create_lending_pool( who: &T::AccountId, id: LendingPoolId, asset: AssetIdOf<T>, balance: BalanceOf<T>, ) -> DispatchResult

pub fn do_activate_lending_pool(asset: AssetIdOf<T>) -> DispatchResult

pub fn do_supply( who: &T::AccountId, asset: AssetIdOf<T>, balance: BalanceOf<T>, ) -> DispatchResult

pub fn do_withdrawal( who: &T::AccountId, asset: AssetIdOf<T>, balance: BalanceOf<T>, ) -> DispatchResult

This method allows a user to withdraw liquidity from a lending pool. The pool can be deactivated or not, but the user must have enough LP tokens to withdraw. This method withdraw some liquidity from a liquidy pool and burns LP tokens of the user

pub fn do_deactivate_lending_pool(asset: AssetIdOf<T>) -> DispatchResult

This method de-activates an existing lending pool

pub fn account_id() -> T::AccountId

This method returns the palled account id

This actually does computation. If you need to keep using it, then make sure to cache the value and only call this once.

pub fn update_and_mint( who: &T::AccountId, asset: AssetIdOf<T>, lp_id: AssetIdOf<T>, scaled_balance: AssetBalanceOf<T>, current_supply_index: Rate, ) -> DispatchResult

Calculates the new_mint amount as follows, let interest_on_old_deposit = old_balance * (current_supply_index - last_supply_index) let total_new_mint = interest_on_old_deposit + new balance Mint total_new_mint LP tokens into supplier’s account Update supplier data

impl<T: Config> Pallet<T>

pub fn reserve_pools<KArg>(k: KArg) -> Option<LendingPool<T>>
where KArg: EncodeLike<AssetPool<T>>,

An auto-generated getter for LendingPoolStorage.

impl<T: Config> Pallet<T>

pub fn last_accrued_interest_time<KArg>(k: KArg) -> UnderlyingAsset<T>
where KArg: EncodeLike<AssetIdOf<T>>,

An auto-generated getter for UnderlyingAssetStorage.

impl<T: Config> Pallet<T>

pub fn max_exchange_rate() -> (Rate, Rate)

An auto-generated getter for MinMaxExchangeRate.