Trait lending::pallet::Config

pub trait Config: Config {
    type PalletId: Get<PalletId>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type NativeBalance: Inspect<Self::AccountId> + Mutate<Self::AccountId> + Inspect<Self::AccountId> + Mutate<Self::AccountId> + Inspect<Self::AccountId> + Mutate<Self::AccountId>;
    type Fungibles: Inspect<Self::AccountId, Balance = BalanceOf<Self>, AssetId = u32> + Mutate<Self::AccountId> + Create<Self::AccountId>;
    type WeightInfo: WeightInfo;
    type Time: MomentTime;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type PalletId: Get<PalletId>

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type NativeBalance: Inspect<Self::AccountId> + Mutate<Self::AccountId> + Inspect<Self::AccountId> + Mutate<Self::AccountId> + Inspect<Self::AccountId> + Mutate<Self::AccountId>

Type to access the Balances Pallet.

type Fungibles: Inspect<Self::AccountId, Balance = BalanceOf<Self>, AssetId = u32> + Mutate<Self::AccountId> + Create<Self::AccountId>

Type to access the Assets Pallet.

type WeightInfo: WeightInfo

The origin which can add or remove LendingPools and update LendingPools TODO Weight information for extrinsics in this pallet.

type Time: MomentTime

Object Safety

This trait is not object safe.