Module lending::storage::hashed

Operation on runtime storage using hashed keys.

Functions

exists

Check to see if key has an explicit entry in storage.

get

Return the value of the item in storage under key, or None if there is no explicit entry.

get_or

Return the value of the item in storage under key, or default_value if there is no explicit entry.

get_or_default

Return the value of the item in storage under key, or the type’s default if there is no explicit entry.

get_or_else

Return the value of the item in storage under key, or default_value() if there is no explicit entry.

get_raw

Get a Vec of bytes from storage.

kill

Ensure key has no explicit entry in storage.

put

Put value in storage under key.

put_raw

Put a raw byte slice into storage.

take

Remove key from storage, returning its value if it had an explicit entry or None otherwise.

take_or

Return the value of the item in storage under key, or default_value if there is no explicit entry. Ensure there is no explicit entry on return.

take_or_default

Remove key from storage, returning its value, or, if there was no explicit entry in storage, the default for its type.

take_or_else

Return the value of the item in storage under key, or default_value() if there is no explicit entry. Ensure there is no explicit entry on return.