Trait lending::storage::types::TupleToEncodedIter

pub trait TupleToEncodedIter {
    // Required method
    fn to_encoded_iter(&self) -> IntoIter<Vec<u8>>;
}

Trait to indicate that a tuple can be converted into an iterator of a vector of encoded bytes.

Required Methods

fn to_encoded_iter(&self) -> IntoIter<Vec<u8>>