Trait lending::GetStorageVersion

pub trait GetStorageVersion {
    type CurrentStorageVersion;

    // Required methods
    fn current_storage_version() -> Self::CurrentStorageVersion;
    fn on_chain_storage_version() -> StorageVersion;
}

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case Provides information about the storage version of a pallet.

It differentiates between current and on-chain storage version. Both should be only out of sync when a new runtime upgrade was applied and the runtime migrations did not yet executed. Otherwise it means that the pallet works with an unsupported storage version and unforeseen stuff can happen.

The current storage version is the version of the pallet as supported at runtime. The active storage version is the version of the pallet in the storage.

It is required to update the on-chain storage version manually when a migration was applied.

Required Associated Types

type CurrentStorageVersion

This will be filled out by the pallet macro.

If the storage_version attribute isn’t given this is set to [NoStorageVersionSet] to inform the user that the attribute is missing. This should prevent that the user forgets to set a storage version when required. However, this will only work when the user actually tries to call Self::current_storage_version to compare it against the Self::on_chain_storage_version. If the attribute is given, this will be set to StorageVersion.

Required Methods

fn current_storage_version() -> Self::CurrentStorageVersion

Returns the current storage version as supported by the pallet.

fn on_chain_storage_version() -> StorageVersion

Returns the on-chain storage version of the pallet as stored in the storage.

Object Safety

This trait is not object safe.