Trait lending::IsType

pub trait IsType<T>: Into<T> + From<T> {
    // Required methods
    fn from_ref(t: &T) -> &Self;
    fn into_ref(&self) -> &T;
    fn from_mut(t: &mut T) -> &mut Self;
    fn into_mut(&mut self) -> &mut T;
}

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case Trait to be used when types are exactly same.

This allow to convert back and forth from type, a reference and a mutable reference.

Required Methods

fn from_ref(t: &T) -> &Self

Cast reference.

fn into_ref(&self) -> &T

Cast reference.

fn from_mut(t: &mut T) -> &mut Self

Cast mutable reference.

fn into_mut(&mut self) -> &mut T

Cast mutable reference.

Object Safety

This trait is not object safe.