Trait lending::MaxEncodedLen

pub trait MaxEncodedLen: Encode {
    // Required method
    fn max_encoded_len() -> usize;
}

! # The Lending pallet of Kylix ! ! ## Overview ! ! The Lending pallet is responsible for managing the lending pools and the treasury operations. ! ! The lending pallet adopts a protocol similar to Compound V2 for its lending operations, ! offering a pool-based approach to aggregate assets from all users. ! ! Interest rates adjust dynamically in response to the supply and demand conditions. ! Additionally, for every lending positions a new token is minted, thus enabling the transfer of ! ownership. ! ! Implemented Extrinsics: ! ! 0. create_lending_pool() ! 1. activate_lending_pool() ! 2. supply() ! 3. withdraw() ! 4. borrow() ! 5. repay() ! 6. claim_rewards() ! 7. deactivate_lending_pool() ! 8. update_pool_rate_model() ! 9. update_pool_kink() !

TODO:

rename the pallet to lending and the module to lending

implement the ManagerOrigin type for reserve pool special operations

implement tests for the lending logic

implement the WeightInfo trait for the pallet

! Use case Items implementing MaxEncodedLen have a statically known maximum encoded size.

Some containers, such as BoundedVec, have enforced size limits and this trait can be implemented accurately. Other containers, such as StorageMap, do not have enforced size limits. For those containers, it is necessary to make a documented assumption about the maximum usage, and compute the max encoded length based on that assumption.

Required Methods

fn max_encoded_len() -> usize

Upper bound, in bytes, of the maximum encoded size of this item.

Object Safety

This trait is not object safe.